Differentiate between adding constraints and laying out subtree
Summary: Callers may not want to cause a layout pass when adding constraints; `layout` maintains the previous behavior, but calling `constrain` opts out of calling layoutIfNeeded/layoutSubtreeIfNeeded.

Test Plan: Add constraints to a view in `init` (before it's added to a window), and notice the lack of 'unsatisfiable constraints' warnings during the layout pass.

Reviewers: mgp, andy

Differential Revision: http://phabricator.khanacademy.org/D14045