Improving edges insetting support
This allows a more expressive use of edges insetting, as it moves the
current function to be a member of `Edges` itself.

e.g. `view.edges == inset(view.superview!.edges, 10)` can now be
written as `view.edges == view.superview!.edges.inseted(by: 10)` which
is a lot more expressive

This commit does not break the current implementation of the `inset`
function, only adds this cosmetic change that allows more expressivity